+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @cosmicdrift/kumiko-cli
 
+## 0.2.311
+
+### Patch Changes
+
+- @cosmicdrift/kumiko-dev-server@0.209.1
+
+## 0.2.310
+
+### Patch Changes
+
+- @cosmicdrift/kumiko-dev-server@0.209.0
+
 ## 0.2.309
 
 ### Patch Changes