+++ packages/dispatcher-live/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @cosmicdrift/kumiko-dispatcher-live
 
+## 0.209.1
+
+### Patch Changes
+
+- @cosmicdrift/kumiko-headless@0.209.1
+
+## 0.209.0
+
+### Patch Changes
+
+- @cosmicdrift/kumiko-headless@0.209.0
+
 ## 0.208.3
 
 ### Patch Changes